requirements.txt
streamlit
openai
azure-ai-documentintelligence
azure-core
python-dotenv